requirements.txt
streamlit
langchain
openai
faiss-cpu
boto3
python-docx
